04_web-faas-builder/pyproject.toml
[project]
name = "blue-faas"
version = "0.1.0"
description = "FastAPI-based Spin application deployment tool for Kubernetes"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "faker>=22.0.0",
    "python-multipart>=0.0.6",
    "boto3>=1.34.0",
    "httpx>=0.26.0",
    "mypy>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.92.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"

[tool.hypothesis]
max_examples = 100

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
